requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.6
django-ninja==1.4.1
pillow==11.1.0
pydantic==2.11.3
pydantic_core==2.33.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
